refactor: move src/providers/ to src/agents/
Reflect the new naming: each module here is an agent (an AI validator),
not a generic provider. registry.ts updates its import paths to match.
