+++ src/adapters/registry.ts
@@ -1,7 +1,7 @@
 import type { Agent } from '../rule.js'
-import { claudeAgentSdk, type QueryFn } from '../providers/claude-agent-sdk.js'
-import { codexSdk } from '../providers/codex-sdk.js'
-import { githubCopilotSdk } from '../providers/github-copilot-sdk.js'
+import { claudeAgentSdk, type QueryFn } from '../agents/claude-agent-sdk.js'
+import { codexSdk } from '../agents/codex-sdk.js'
+import { githubCopilotSdk } from '../agents/github-copilot-sdk.js'
 import type { Adapter } from './adapter.js'
 import * as claudeCode from './claude-code.js'
 import * as codex from './codex.js'
