+++ src/agents/ai-client-from-text.test.ts
