+++ src/agents/ai-client-from-text.ts
