+++ src/agents/parse-verdict.ts
